tweak names and get controller specs passing
