+++ app/controllers/external_users/advocates/claims_controller.rb
@@ -5,7 +5,6 @@ class ExternalUsers::Advocates::ClaimsController < ExternalUsers::ClaimsControll
   before_action :load_advocates_in_provider, only: [:new, :edit, :create, :update]
 
   def new
-    # ap "#{self.class}: NEW"
     @claim = Claim::AdvocateClaim.new
     load_offences_and_case_types
     build_nested_resources
@@ -17,13 +16,12 @@ class ExternalUsers::Advocates::ClaimsController < ExternalUsers::ClaimsControll
       create_and_submit
     else
       create_draft
-  end
+    end
   end
 
 private
 
   def load_advocates_in_provider
-    # ap "#{self.class}: LOAD_ADVOCATES_IN_PROVIDER"
     @advocates_in_provider = @provider.advocates if @external_user.admin? && self.class == ExternalUsers::Advocates::ClaimsController
   end
