+++ app/views/external_users/claim_types/index.html.haml
