+++ app/views/external_users/claims/_form.html.haml
@@ -1,7 +1,6 @@
 #claim-form
   = form_for [:external_users, @claim], as: :claim, url: url_for_external_users_claim(@claim), builder: AdpFormBuilder, html: { novalidate: true }, multipart: true, authenticity_token: true do |f|
 
-    = f.hidden_field :claim_class, value: @claim_class
     = f.hidden_field :form_id, value: @form_id || params[:claim][:form_id]
 
     = render partial: 'external_users/claims/case_details/fields', locals: { f: f }
