+++ app/views/external_users/claims/confirmation.html.haml
@@ -19,4 +19,4 @@
   .button-holder
     = link_to t('link.view_claim'), external_users_root_url, class: 'button button-secondary'
 
-    = link_to t('link.file_another_claim'), external_users_claims_options_path, class: 'button'
+    = link_to t('link.file_another_claim'), types_external_users_claims_path, class: 'button'
