+++ spec/factories/external_users.rb
@@ -57,6 +57,7 @@ FactoryGirl.define do
       roles ['admin']
       provider { create(:provider, :agfs_lgfs) }
     end
+
   end
 end
